refactor(coordinator): extract ConflationAppOrchestrator from CoordinatorApp
Moves ForcedTransactionsApp instantiation (including RISC-V cutover logic),
ConflationAppV1, and ConflationAppV2 lifecycle management into a new
ConflationAppOrchestrator class. CoordinatorApp now only holds coordinator-wide
infrastructure and delegates conflation lifecycle to the orchestrator.

The orchestrator is the natural home for the upcoming resume block number
resolution and DB cleanup on V1→V2 cutover.
